# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Undercore Orchestrator — backend API",
    "version": "0.4.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "description": "New integrations: v2 plus process health and node overview. Legacy v1 client operations are documented separately in docs/public/api.md. Trusted backend only; it authorizes users and slots. Bodies are limited to 8192 bytes. Runtime validation includes identity, expiry and state checks beyond JSON Schema. HTTPS is required off loopback. This artifact never enables live Swagger."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8792",
      "description": "Local development only"
    }
  ],
  "security": [
    {
      "backendToken": []
    }
  ],
  "paths": {
...
    "/internal/v1/nodes": {
      "get": {
        "operationId": "list_nodes",
        "summary": "List nodes",
        "description": "Registry and assignment metadata, not live speed tests.",
        "responses": {
          "200": {
            "description": "Successful operation. Tunnel connectivity is not implied.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "private, no-store"
              },
              "X-Request-ID": {
                "description": "Server-generated correlation ID; caller input is not trusted",
                "schema": {
                  "type": "string",
                  "pattern": "^[a-f0-9]{32}$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Nodes"
                }
              }
            }
          },
          "default": {
            "description": "Application error: detail is a machine code (contracts/errors.json). 401 authorization; 404 ownership/not found; 409 conflict; 410 unavailable; 413/422 invalid input; 429 cooldown; 503 unavailable or pending lease. Framework query-validation errors can instead contain a detail array. Proxies may return non-JSON errors. Repeat only the same operation identity.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "detail": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      },
                      "required": [
                        "detail"
                      ]
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation ID; caller input is not trusted",
                "schema": {
                  "type": "string",
                  "pattern": "^[a-f0-9]{32}$"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "backendToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ]
      },
...
      "Nodes": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "id": {
                  "type": "string"
                },
                "region": {
                  "type": "string"
                },
                "mode": {
                  "enum": [
                    "active",
                    "draining",
                    "disabled"
                  ]
                },
                "assigned": {
                  "type": "integer"
                },
                "capacity": {
                  "type": "integer"
                },
                "pending": {
                  "type": "integer"
                },
                "last_operation": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "last_outcome": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "checked_at": {
                  "type": [
                    "number",
                    "null"
                  ]
                }
              },
              "required": [
                "id",
                "region",
                "mode",
                "assigned",
                "capacity",
                "pending",
                "last_operation",
                "last_outcome",
                "checked_at"
              ]
            }
          }
        },
        "required": [
          "nodes"
        ]
      }
    }
  }
}